requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
numpy==1.26.3
requests==2.31.0
gunicorn==23.0.0
scikit-learn==1.4.2
Pillow==10.2.0
tensorflow==2.12.0